requirements.txt
asgiref==3.8.1
Django==5.1.6
django-allauth==65.4.1
django-htmx==1.22.0
django-simple-history==3.8.0
djangorestframework==3.15.2
sqlparse==0.5.3
